# stage-pfe-front/tools/swagger/src/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Swagger API documentation for Gestion de stock application",
    "version" : "0.0.1-SNAPSHOT",
    "title" : "Swagger API Documentation for Gestion de stock application",
    "contact" : {
      "name" : "Ali Bouali",
      "url" : "https://www.udemy.com/user/bouali-ali/",
      "email" : "bouali.social@gmail.com"
    }
  },
  "basePath" : "http://localhost:8080",
  "paths" : {
...
    "/api/sessions/{id}" : {
      "get" : {
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SessionsDto"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/api/themes/exists" : {
      "get" : {
        "operationId" : "themeExists",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
    "/api/themes/formation/{formationId}" : {
      "get" : {
        "operationId" : "findThemesByFormation",
        "parameters" : [ {
          "name" : "formationId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ThemesDto"
              }
            }
          }
        }
      }
    },
    "/api/themes/search" : {
      "get" : {
        "operationId" : "findByName",
        "parameters" : [ {
          "name" : "name",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ThemesDto"
              }
            }
          }
        }
      }
    },
    "/api/themes/{id}" : {
      "get" : {
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ThemesDto"
            }
          }
        }
      },
      "put" : {
        "operationId" : "updateTheme",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/ThemesDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ThemesDto"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/api/themes/{themeId}/formation/{formationId}" : {
      "post" : {
        "operationId" : "assignThemeToFormation",
        "parameters" : [ {
          "name" : "themeId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "formationId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ThemesDto"
            }
          }
        }
      }
    },
...
    "/certificats" : {
      "get" : {
        "operationId" : "getAllCertificats",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/CertificatsDto"
              }
            }
          }
        }
      },
      "post" : {
        "operationId" : "genererCertificat",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/CertificatsDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CertificatsDto"
            }
          }
        }
      }
    },
    "/certificats/{idUtilisateur}" : {
      "get" : {
        "operationId" : "getCertificatById",
        "parameters" : [ {
          "name" : "idUtilisateur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CertificatsDto"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "supprimerCertificat",
        "parameters" : [ {
          "name" : "idUtilisateur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/save" : {
      "post" : {
        "operationId" : "save",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/SallesDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SallesDto"
            }
          }
        }
      }
    },
    "/{id}" : {
      "get" : {
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SallesDto"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    }
  },
  "definitions" : {
    "CertificatsDto" : {
      "type" : "object",
      "properties" : {
        "inscriptionId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dateGeneration" : {
          "type" : "string",
          "format" : "date"
        },
        "numeroSerie" : {
          "type" : "string"
        },
        "statut" : {
          "type" : "string"
        }
      }
    },
...
    "SallesDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "capacite" : {
          "type" : "integer",
          "format" : "int32"
        },
        "equipement" : {
          "type" : "string"
        },
        "statut" : {
          "type" : "string"
        }
      }
    },
    "SessionsDto" : {
      "type" : "object",
      "properties" : {
        "sessionId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "formationId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "utilisateursIds" : {
          "type" : "array",
          "items" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "salleId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dateDebut" : {
          "type" : "string",
          "format" : "date"
        },
        "dateFin" : {
          "type" : "string",
          "format" : "date"
        },
        "capacite" : {
          "type" : "integer",
          "format" : "int32"
        },
        "placesDisponibles" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "ThemesDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "formationsIds" : {
          "type" : "array",
          "items" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      }
    },
...
  }
}